students/k33391/Igor_Tulukin/Lr3/task2/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
attrs==24.2.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
docker==7.1.0
ecdsa==0.19.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
multidict==6.1.0
orjson==3.9.14
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.0
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-i18n==0.4.4
pydantic-settings==2.2.1
pydantic_core==2.16.2
PyJWT==2.8.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.30
sqlmodel==0.0.16
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
urllib3==2.2.3
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.12.1
